[0.6.5] remove chunkrenderer (it's buggy)

+++ include/cube/graphics/Vertex.hpp
@@ -14,8 +14,8 @@ namespace cube {
     };
 
     struct Vertex3D {
-        glm::u16vec3 pos;
-        glm::u16vec2 tex;
+        glm::vec3 pos;
+        glm::vec2 tex;
     };
 }
